Overhaul of template.ssjs
======== THERE IS NEW SYNTAX ========
It is absolutely forbidden to include one tag inside another... NO
execptions anymore.  This one will have the largest impact.  Essentially,
@@JS:if(somevar)then'@@something:else@@'else''@@ will *NOT* work.
Now, if you want to access to current iteration of the repeated array
inside a <<REPEAT objname>> section, it's available inside of JS: bits as
RepeatObj.  This should remove all need for nested tags.

A new convinience functions has been added... Nz(value[, valIfUndef])
returns valIfUndef (defaults to the empty string) if value is undefined
and value if it IS defined.