Start of translation to JS. Converted all the level/flags/age restrictions except the minor cutoff to ARS strings.
NOTE: JavaScript currently has no way of knowing what directory the script
in that I can tell.  This means that we can't reliably have stand-alone JS
doors easily as the datafiles etc would need to be in a known location.

Simply making the script path available to JS would solve this issue, but a
quick dig through the source indicates that this would be relatively
difficult.