Add an 'x' command to end a turn and a warpoll.js script.
When all players mark themselves as "turn done" then the next time warpoll.js
runs, it will run the warupd.js and advance the game by one turn.  Polled
updates add a flag to the save game indicating it was last updated by a poll.
During regular maintenance, if the last update was a polled update, the
update only clears the flag and does not advance the game.