Sent Shift In (\x0f) after sending the test music sequence.
ANSI music terminates with a \x0e (Shift Out) which, for terminals that
support it, switches to the G1 character set.  This change switches back
to the G0 character set (and deletes the white sun with rays character) to
avoid leaving the console in the G1 set.

Should fix problems reported with KnK on the Linux console.