Update Cryptlib to 3.4.5
This versions has a lot more assumptions that the whole world runs Linux
in it, and just a small number of perviously reported bugs actually fixed.

I think the patch count has actually gone up rather than down for this.

As it turns out, the thing I was trying to fix (https://pskreporter.info not
working with HTTP.js) is not fixed in the new version and was just yet
another abitrary undocumented limit being hit.

This *may* break anything crypto-releated, though it seems to work on FreeBSD
fine.  Win32 build not tested since Digital Man provides precompiled libs
for that, and mingw32 builds not yet tested.

Please report any *NEW* crypto-related issues after getting this to work.

Oh yeah, it also looks like the cryptlib dependency for Synchronet got
broken somewhere, so the easiest way to rebuild with the new cryptlib is
to build SyncTERM first.  I'll try to get this one fixed soonish.