Attempt to fix issues on x86 systems by driving a final stake through all inline asm stuff (cl-noasm-defines.patch).
This (of course) revealed an issue with 64-bit systems not building properly
so cl-bn-noasm64-fix.patch fixes this again.

There should no longer be any inline (or out of line) asm, so any more
illegal instruction errors are likely to be the fault of the compiler or
the compiler flags.